Split diesel middleware into its own module
